extract out rc file management + fix newline bug
