+++ shell-setup/src/main.ts
@@ -1,5 +1,5 @@
 import { environment } from "./environment.ts";
-import { basename, dirname, join } from "@std/path";
+import { dirname, join } from "@std/path";
 import { confirm, multiSelect } from "@nathanwhit/promptly";
 import { parseArgs } from "@std/cli/parse-args";
 
@@ -11,17 +11,10 @@ import {
   shEnvScript,
   shSourceString,
   type UnixShell,
-  type UpdateRcFile,
   Zsh,
 } from "./shell.ts";
-import {
-  ensureEndsWith,
-  ensureExists,
-  ensureStartsWith,
-  info,
-  warn,
-  withContext,
-} from "./util.ts";
+import { ensureExists, warn } from "./util.ts";
+import { RcBackups, updateRcFile } from "./rc_files.ts";
 const {
   readTextFile,
   runCmd,
@@ -96,28 +89,10 @@ async function writeCompletionFiles(
   return results;
 }
 
-/** A little class to manage backing up shell rc files */
-class Backups {
-  backedUp = new Set<string>();
-  constructor(public backupDir: string) {}
-
-  async add(path: string, contents: string): Promise<void> {
-    if (this.backedUp.has(path)) {
-      return;
-    }
-    const dest = join(this.backupDir, basename(path)) + `.bak`;
-    info(
-      `backing '${path}' up to '${dest}'`,
-    );
-    await Deno.writeTextFile(dest, contents);
-    this.backedUp.add(path);
-  }
-}
-
 /** Write commands necessary to set up completions to shell rc files */
 async function writeCompletionRcCommands(
   availableShells: UnixShell[],
-  backups: Backups,
+  backups: RcBackups,
 ) {
   for (const shell of availableShells) {
     if (!shell.supportsCompletion) continue;
@@ -152,88 +127,13 @@ async function writeEnvFiles(availableShells: UnixShell[], installDir: string) {
   }
 }
 
-/** Updates an rc file (e.g. `.bashrc`) with a command string.
- * If the file already contains the command, it will not be updated.
- * @param rc - path to the rc file
- * @param command - either the command to append, or an object with commands to prepend and/or append
- * @param backups - manager for rc file backups
- */
-async function updateRcFile(
-  rc: string,
-  command: string | UpdateRcFile,
-  backups: Backups,
-): Promise<boolean> {
-  let prepend = "";
-  let append = "";
-  if (typeof command === "string") {
-    append = command;
-  } else {
-    prepend = command.prepend ?? "";
-    append = command.append ?? "";
-  }
-  if (!prepend && !append) {
-    return false;
-  }
-
-  let contents: string | undefined;
-  try {
-    contents = await readTextFile(rc);
-    if (prepend) {
-      if (contents.includes(prepend)) {
-        // nothing to prepend
-        prepend = "";
-      } else {
-        // always add a newline
-        prepend = ensureEndsWith(prepend, "\n");
-      }
-    }
-    if (append) {
-      if (contents.includes(append)) {
-        // nothing to append
-        append = "";
-      } else if (!contents.endsWith("\n")) {
-        // add new line to start
-        append = ensureStartsWith(append, "\n");
-      }
-    }
-  } catch (_error) {
-    prepend = prepend ? ensureEndsWith(prepend, "\n") : prepend;
-  }
-  if (!prepend && !append) {
-    return false;
-  }
-
-  if (contents !== undefined) {
-    await backups.add(rc, contents);
-  }
-
-  await ensureExists(dirname(rc));
-
-  try {
-    await writeTextFile(rc, prepend + (contents ?? "") + append, {
-      create: true,
-    });
-
-    return true;
-  } catch (error) {
-    if (
-      error instanceof Deno.errors.PermissionDenied ||
-      // deno-lint-ignore no-explicit-any
-      error instanceof (Deno.errors as any).NotCapable
-    ) {
-      return false;
-    }
-    throw withContext(`Failed to update shell rc file: ${rc}`, error);
-  }
-}
-
 /** Write the commands necessary to source the env file (which sets up the path).
  * Up until this point, we have not modified any shell config files.
  */
 async function addToPath(
   availableShells: UnixShell[],
   installDir: string,
-  backups: Backups,
+  backups: RcBackups,
 ) {
   for (const shell of availableShells) {
     const sourceCmd = await (shell.sourceString ?? shSourceString)(installDir);
@@ -284,7 +184,7 @@ async function setupShells(
 
   await writeEnvFiles(availableShells, installDir);
 
-  const backups = new Backups(backupDir);
+  const backups = new RcBackups(backupDir);
 
   if (
     (skipPrompts && !noModifyPath) || (!skipPrompts &&
